services/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic[email]==2.5.0
httpx==0.25.1
PyJWT==2.8.1
python-multipart==0.0.6
python-dateutil==2.8.2
psutil==5.9.6
# Optional: For production database support
sqlalchemy==2.0.23
# Optional: For async database support
asyncpg==0.29.0
# Optional: For monitoring and tracing
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
# Optional: For distributed logging
python-json-logger==2.0.7
